domain92/requirements_discord.txt
discord.py>=2.3.0
pillow>=11.1.0
requests>=2.32.3
art>=6.4
freedns-client>=0.1.3
pytesseract>=0.3.13
lolpython>=2.2.0
pysocks>=1.7.1
stem>=1.8.2 